Restructured the app for better group collab
