+++ app/(tabs)/ask.tsx
@@ -1,0 +1,1 @@
+export { default } from '@screens/ask';
