+++ app/request/[id].tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { useLocalSearchParams } from "expo-router";
+import { View, Text, StyleSheet } from "react-native";
+
+export default function RequestDetailScreen() {
+  const { id } = useLocalSearchParams();
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Request #{id}</Text>
+      <Text>This is where request details will appear.</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: { flex: 1, padding: 20, backgroundColor: "#fff" },
+  title: { fontSize: 24, fontWeight: "bold", marginBottom: 10 },
+});
