+++ app/screens/diploma.tsx
