+++ components/ui/card.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function Card({ children, className }: { children: React.ReactNode; className?: string }) {
+  return <div className={`rounded-lg border ${className ?? ""}`}>{children}</div>;
+}
+
+export function CardContent({ children, className }: { children: React.ReactNode; className?: string }) {
+  return <div className={className}>{children}</div>;
+}
