+++ constants/Colors.ts
@@ -1,19 +1,13 @@
-const tintColorLight = '#2f95dc';
-const tintColorDark = '#fff';
-
-export default {
-  light: {
-    text: '#000',
-    background: '#fff',
-    tint: tintColorLight,
-    tabIconDefault: '#ccc',
-    tabIconSelected: tintColorLight,
-  },
-  dark: {
-    text: '#fff',
-    background: '#000',
-    tint: tintColorDark,
-    tabIconDefault: '#ccc',
-    tabIconSelected: tintColorDark,
-  },
+const Colors = {
+  primary: "#A6192E",
+  background: "#FFFFFF",
+  text: "#000000",
+  gray: "#666666",
+  border: "#CCCCCC",
+  tag: "#005EB8",
+  card: "#f6f7f9",
+  muted: "#999999",
 };
+
+export default Colors;
+
